requirements.txt
# Modern Alpaca SDK
alpaca-py>=0.15.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Async HTTP requests
aiohttp>=3.8.0
asyncio

# Google Cloud services
google-cloud-storage>=2.10.0
google-auth>=2.15.0

# Date/time handling
pytz>=2023.3
holidays>=0.35

# Technical analysis
TA-Lib>=0.4.25

# Web requests
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0

# Data validation
pydantic>=2.0.0

# Mathematical operations
scipy>=1.11.0

# JSON handling
ujson>=5.8.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
